--- App.jsx
@@ -1,1 +1,0 @@
-export default () => '🔥 Ready.'
