+++ chrome/src/containers/SingleHighlight.jsx
@@ -6,22 +6,9 @@ import Interactive from '../components/Interactive';
 import CreateHighlightButton from '../components/CreateHighlightButton';
 import { urlEncode } from '../highlighting';
 
-export default class HighlightAnnotations extends Component {
+export default class SingleHighlight extends Component {
   constructor(props) {
     super(props);
-
-    //get the active highlight (most recently clicked)
-    // const selected = document.getElementsByClassName('activeHighlight')[0];
-    // get an array of classes on it
-    // const selectedClasses = selected.className.split(' ');
-    // const selectedId = selectedClasses.filter(el => {
-    //   return el !== 'activeHighlight' && el !== 'chromelights-highlights';
-    // });
-
-    /* the selectedId should be used to find and render all of the
-         annotations associated with the selected highlight */
-
-    // console.log('classes', selectedId);
 
     this.state = {
       selectedHighlight: this.props.activeId || 'Select Some Text'
