--- lib/index.js
@@ -1,5 +1,0 @@
-export default 'hello there'
-export const description = `This library is shared with your firebase functions.
-
-It's bundled by webpack into ~/functions/lib.js when you npm start.
-`
