--- public/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-  </head>
-  <body>
-    <div id="main">
-      <h1>Hello Chromies!</h1>
-    </div>
-    <script async defer src="/bundle.js"></script>
-  </body>
-</html>
