--- test-highlight.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Document</title>
-  <style type="text/css">
-    #main {
-      margin: 40px;
-    }
-
-    ol {
-      margin: 10px 40px
-    }
-  </style>
-</head>
-
-<body>
-  <div id="main">
-    <h1>Header</h1>
-    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quidem reprehenderit doloribus, cupiditate corrupti illum vitae
-      numquam ex facilis sequi ipsum, optio iusto, at expedita quasi itaque voluptatibus molestias dolorem fugiat!</p>
-
-    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quidem reprehenderit doloribus, cupiditate corrupti illum vitae
-      numquam ex facilis sequi ipsum, optio iusto, at expedita quasi itaque voluptatibus molestias dolorem fugiat!</p>
-
-    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quidem reprehenderit doloribus, cupiditate corrupti illum vitae
-      numquam ex facilis sequi ipsum, optio iusto, at expedita quasi itaque voluptatibus molestias dolorem fugiat!
-    </p>
-    <div>
-      <h2>Subheader</h2>
-      <div>
-        <ol>
-          <li>Lorem ipsum</li>
-          <li>Lorem ipsum</li>
-          <li>Lorem ipsum</li>
-          <li>Lorem ipsum</li>
-        </ol>
-
-      </div>
-    </div>
-    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quidem reprehenderit doloribus, cupiditate corrupti illum vitae
-      numquam ex facilis sequi ipsum, optio iusto, at expedita quasi itaque voluptatibus molestias dolorem fugiat!
-    </p>
-
-
-  </div>
-</body>
-</body>
-
-</html>
